✨ feat(add dataset change pre): csv->txt->list pre
In order to meet the needs of pre-training, the data needs to be converted into a list, which contains elements with a length of 512. So I chose to convert the original csv file into a txt file, and add a space every 512 bp as a separator. When creating the dataset, just read the corresponding txt file directly.

I'm not sure if this method is a general method, and I'm not sure if it will lead to a reduction in efficiency.